start import metadata tests
